refactor: move UserRole enum to a separate file
