+++ src/modules/users/enum/roles.ts
@@ -1,0 +1,6 @@
+export enum UserRole {
+  ADMIN = 'ROLE_ADMIN',
+  GESTOR = 'ROLE_GESTOR',
+  AGENTE = 'ROLE_AGENTE',
+  PUBLIC = 'PUBLIC'
+}
